New Generate Audiobook tab added. -utilized most code from ebook2audiobookSTYLETTS2  with a little cleanup and condensing but the core  functionality of converting ebooks to txt files  is pretty much just copied and pasted. -added some logic to parse through the generated  .txt files to clean up the formatting. -it will change the beta by +/-2 depending for  anything in the ebook wrapped in quotes to be  able to better distinguish while listening.
The audiobook generator is very crudely coded,
rushed to get a working proof of concept. Much
code cleanup need to be done, and I assume there
will always be the occasional ebook file that won't
generate.

As I have time to go over the code I'll clean it up,
but feel free to do so yourself if you feel froggy.

Always consider that the audiobook isn't going to
be perfect and is HEAVILY dependant on the model
you use and how well it's been trained. I trained
a model using a lot of audiobook clips narrated
by Julia Whelan and it's really good, but there are
definitely things that can be done to improve a
bit here and there.

I still need to update the requirements.txt for
some things used by the ebook2audiobookSTYLETTS2,
so this won't work for you out of the box without
a number of manual pip installs, fyi.