New history tab and ID3 tagging, cleanup Implemented mutagen to tag all generated files with a bunch of data related to the generation settings etc. The structure of the generated files is now ./results/[voice]/[voice][date formatted filename-safe down to the second].wav
As of this writing, the following ID3 tags are created on each tile (txxxx)
-seed
-original_seed
-alpha
-beta
-diffusion_steps
-embedding_scale
-reference_audio_path
-voice
-voice_model
-rtf
-text
-date_generated

Using the above ID3 tags, the new history tab will list all files in the results folder (and subfolders) generated by the app and list them on the history tab.
Clicking a result in the dataframe on the history tab will populated all the current ID3 tag information in one block, the text used to generate it in another block, and an audio block to play the file.
-todo: implement "send to generation tab" type functionality as seen in many other generative gradio apps.